Refactor resource managers
Move resource manager in a separate file.
